+++ src/AppInstallerCommonCore/Public/AppInstallerErrors.h
@@ -128,15 +128,16 @@
 #define APPINSTALLER_CLI_ERROR_INSTALL_DEPENDENCIES             ((HRESULT)0x8A150110)
 
 // Status values for check package installed status results.
+// Partial success has the success bit(first bit) set to 0.
 #define WINGET_INSTALLED_STATUS_ARP_ENTRY_FOUND                             S_OK
 #define WINGET_INSTALLED_STATUS_ARP_ENTRY_NOT_FOUND                         ((HRESULT)0x8A150201)
 #define WINGET_INSTALLED_STATUS_INSTALL_LOCATION_FOUND                      S_OK
-#define WINGET_INSTALLED_STATUS_INSTALL_LOCATION_NOT_APPLICABLE             ((HRESULT)0x8A150202)
+#define WINGET_INSTALLED_STATUS_INSTALL_LOCATION_NOT_APPLICABLE             ((HRESULT)0x0A150202)
 #define WINGET_INSTALLED_STATUS_INSTALL_LOCATION_NOT_FOUND                  ((HRESULT)0x8A150203)
 #define WINGET_INSTALLED_STATUS_FILE_HASH_MATCH                             S_OK
 #define WINGET_INSTALLED_STATUS_FILE_HASH_MISMATCH                          ((HRESULT)0x8A150204)
 #define WINGET_INSTALLED_STATUS_FILE_NOT_FOUND                              ((HRESULT)0x8A150205)
-#define WINGET_INSTALLED_STATUS_FILE_FOUND_WITHOUT_HASH_CHECK               ((HRESULT)0x8A150206)
+#define WINGET_INSTALLED_STATUS_FILE_FOUND_WITHOUT_HASH_CHECK               ((HRESULT)0x0A150206)
 #define WINGET_INSTALLED_STATUS_FILE_ACCESS_ERROR                           ((HRESULT)0x8A150207)
