+++ src/AppInstallerRepositoryCore/AppInstallerRepositoryCore.vcxproj
@@ -354,6 +354,7 @@
     <ClCompile Include="Microsoft\SQLiteIndex.cpp" />
     <ClCompile Include="Microsoft\SQLiteIndexSource.cpp" />
     <ClCompile Include="PackageDependenciesValidation.cpp" />
+    <ClCompile Include="PackageInstalledStatus.cpp" />
     <ClCompile Include="PackageTrackingCatalog.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
