+++ src/AppInstallerRepositoryCore/Rest/RestSource.cpp
@@ -85,11 +85,6 @@ namespace AppInstaller::Repository::Rest
                 return false;
             }
 
-            std::vector<InstallerInstalledStatus> CheckInstalledStatus(InstalledStatusType) const override
-            {
-                THROW_HR(E_NOTIMPL);
-            }
-
             bool IsSame(const IPackage* other) const override
             {
                 const AvailablePackage* otherAvailablePackage = dynamic_cast<const AvailablePackage*>(other);
